+++ bin/check-cloudwatch-composite-metric.rb
@@ -85,14 +85,14 @@ class CloudWatchCompositeMetricCheck < Sensu::Plugin::Check::CLI
          description: 'CloudWatch metric unit'
 
   option :critical,
-         description: 'Trigger a critical when value is over VALUE',
+         description: 'Trigger a critical when value is over VALUE as a Percent',
          short: '-c VALUE',
          long: '--critical VALUE',
          proc: proc(&:to_f),
          required: true
 
   option :warning,
-         description: 'Trigger a warning when value is over VALUE',
+         description: 'Trigger a warning when value is over VALUE as a Percent',
          short: '-w VALUE',
          long: '--warning VALUE',
          proc: proc(&:to_f)
@@ -103,20 +103,97 @@ class CloudWatchCompositeMetricCheck < Sensu::Plugin::Check::CLI
          long: '--operator OPERATION',
          default: 'greater'
 
-  option :no_data_ok,
-         short: '-O',
-         long: '--allow-no-data',
-         description: 'Returns ok if no data is returned from the metric',
+  option :numerator_default,
+         long: '--numerator-default DEFAULT',
+         description: 'Default for numerator if no data is returned for metric',
+         proc: proc(&:to_f)
+
+  option :no_denominator_data_ok,
+         long: '--allow-no-denominator-data',
+         description: 'Returns ok if no data is returned from denominator metric',
          boolean: true,
          default: false
 
+  option :zero_denominator_data_ok,
+         long: '--allow-zero-denominator-data',
+         description: 'Returns ok if denominator metric is zero',
+         boolean: true,
+         default: false
+
+  option :no_data_ok,
+         short: '-O',
+         long: '--allow-no-data',
+         description: 'Returns ok if no data is returned from either metric',
+         boolean: true,
+         default: false
   include CloudwatchCommon
 
   def metric_desc
     "#{config[:namespace]}-#{config[:numerator_metric_name]}/#{config[:denominator_metric_name]}(#{dimension_string})"
   end
 
+  def numerator_data(metric_payload)
+    if resp_has_no_data(metric_payload, config[:statistics])
+      # If the numerator response has no data in it, see if there was a predefined default.
+      # If there is no predefined default it will return nil
+      config[:numerator_default]
+    else
+      read_value(metric_payload, config[:statistics]).to_f
+    end
+  end
+
+  # rubocop:disable Style/GuardClause
+  def composite_check
+    numerator_metric_resp = get_metric(config[:numerator_metric_name])
+    denominator_metric_resp = get_metric(config[:denominator_metric_name])
+
+    ## If the numerator is empty, then we see if there is a default. If there is a default
+    ## then we will pretend the numerator _isnt_ empty. That is
+    ## if empty but there is no default this will be true. If it is empty and there is a default
+    ## this will be false (i.e. there is data, following standard of dealing in the negative here)
+    no_num_data = numerator_data(numerator_metric_resp).nil?
+    no_den_data = resp_has_no_data(denominator_metric_resp, config[:statistics])
+    no_data = no_num_data || no_den_data
+
+    # no data in numerator or denominator this is to keep backwards compatibility
+    if no_data && config[:no_data_ok]
+      return :ok, "#{metric_desc} returned no data but that's ok"
+    elsif no_den_data && config[:no_denominator_data_ok]
+      return :ok, "#{config[:denominator_metric_name]} returned no data but that's ok"
+    elsif no_data ## This is legacy case
+      return :unknown, "#{metric_desc} could not be retrieved"
+    end
+
+    ## Now both the denominator and numerator have data (or a valid default)
+    denominator_value = read_value(denominator_metric_resp, config[:statistics]).to_f
+    if denominator_value.zero? && config[:zero_denominator_data_ok]
+      return :ok, "#{metric_desc}: denominator value is zero but that's ok"
+    elsif denominator_value.zero?
+      return :unknown, "#{metric_desc}: denominator value is zero"
+    end
+
+    ## We already checked if this value is nil so we know its not
+    numerator_value = numerator_data(numerator_metric_resp)
+    value = (numerator_value / denominator_value * 100).to_i
+    base_msg = "#{metric_desc} is #{value}: comparison=#{config[:compare]}"
+
+    if compare(value, config[:critical], config[:compare])
+      return :critical, "#{base_msg} threshold=#{config[:critical]}"
+    elsif config[:warning] && compare(value, config[:warning], config[:compare])
+      return :warning,  "#{base_msg} threshold=#{config[:warning]}"
+    else
+      threshold = config[:warning] || config[:critical]
+      return :ok, "#{base_msg}, will alarm at #{threshold}"
+    end
+  end
+  # rubocop:enable Style/GuardClause
+
   def run
-    composite_check config
+    status, msg = composite_check
+    if respond_to?(status)
+      send(status, msg)
+    else
+      unknown 'unknown exit status called'
+    end
   end
 end