+++ lib/sensu-plugins-aws/cloudwatch-common.rb
@@ -3,7 +3,7 @@ module CloudwatchCommon
   include Common
 
   def client
-    Aws::CloudWatch::Client.new
+    @client ||= Aws::CloudWatch::Client.new
   end
 
   def read_value(resp, stats)
@@ -40,47 +40,13 @@ module CloudwatchCommon
     }
   end
 
-  def composite_metrics_request(config, metric, fixed_time_now = Time.now)
-    {
-      namespace: config[:namespace],
-      metric_name: config[metric],
-      dimensions: config[:dimensions],
-      start_time: fixed_time_now - config[:period] * 10,
-      end_time: fixed_time_now,
-      period: config[:period],
-      statistics: [config[:statistics]],
-      unit: config[:unit]
-    }
+  def get_metric(metric)
+    client.get_metric_statistics(composite_metrics_request(metric))
   end
 
-  def composite_check(config)
-    fixed_time_now = Time.now
-    numerator_metric_resp = client.get_metric_statistics(composite_metrics_request(config, :numerator_metric_name, fixed_time_now))
-    denominator_metric_resp = client.get_metric_statistics(composite_metrics_request(config, :denominator_metric_name, fixed_time_now))
-
-    no_data = resp_has_no_data(numerator_metric_resp, config[:statistics]) || \
-              resp_has_no_data(denominator_metric_resp, config[:statistics])
-    if no_data && config[:no_data_ok]
-      ok "#{metric_desc} returned no data but that's ok"
-    elsif no_data && !config[:no_data_ok]
-      unknown "#{metric_desc} could not be retrieved"
-    end
-
-    denominator_value = read_value(denominator_metric_resp, config[:statistics]).to_f
-    if denominator_value.zero?
-      ok "#{metric_desc} denominator value is zero but that's ok"
-    end
-    numerator_value = read_value(numerator_metric_resp, config[:statistics]).to_f
-    value = (numerator_value / denominator_value * 100).to_i
-    base_msg = "#{metric_desc} is #{value}: comparison=#{config[:compare]}"
-
-    if compare(value, config[:critical], config[:compare])
-      critical "#{base_msg} threshold=#{config[:critical]}"
-    elsif config[:warning] && compare(value, config[:warning], config[:compare])
-      warning "#{base_msg} threshold=#{config[:warning]}"
-    else
-      ok "#{base_msg}, will alarm at #{!config[:warning].nil? ? config[:warning] : config[:critical]}"
-    end
+  def composite_metrics_request(metric)
+    ## config is a class variable but don't want to change signature
+    metrics_request(config).merge(metric_name: metric)
   end
 
   def self.parse_dimensions(dimension_string)